+++ views/timetable.html
@@ -26,6 +26,9 @@
     <div class="flex flex-col col-span-3 sm:col-span-1">
         <span class="text-xs md:text-sm text-blue-200 uppercase opacity-80 font-medium">Departs</span>
         <span class="text-xl font-medium sm:text-2xl md:text-3xl">{{ .DepartureTimeDisplay }} <span class="timeToDepart animate-pulse text-base font-normal" data-zero-time="{{ .OrderTime }}"></span></span>
+        {{ if .X_Platform }}
+            <div class="w-fit text-xs md:text-sm text-blue-100 bg-blue-500 uppercase font-medium px-2 pt-0.5 pb-0 rounded">{{ .X_Platform }}</div>
+        {{ end }}
     </div>
     <div class="flex flex-col col-span-2 sm:col-span-1">
         <span class="text-xs md:text-sm text-green-300 uppercase opacity-80 font-medium">Travels</span>
